Use host-routable IPs in Docker-in-Docker mesos-slave
The Docker-in-Docker docker daemon is reconfigured by using a docker0 bridge
which is bridged against the veth device eth0. This way the inner containers
will live in the same layer 2 network as the outer containers.

In order to avoid IP conflict, --fixed-cidr=a.b.c.d/24 must be passed in the
DOCKER_DAEMON_ARGS env variable, with a IP range a.b.c.d/24 in the outer docker
network range, e.g. 172.17.1.0/24 for a standard docker setup which uses
172.17.42.0/24 for the outer container IPs.